refactor: move classes to more coherent locations
