--- operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/config/Dependent.java
@@ -1,8 +1,0 @@
-package io.javaoperatorsdk.operator.api.config;
-
-public @interface Dependent {
-
-  Class<?> resourceType();
-
-  Class<? extends DependentResource> type();
-}
