+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/config/dependent/Dependent.java
@@ -1,0 +1,10 @@
+package io.javaoperatorsdk.operator.api.config.dependent;
+
+import io.javaoperatorsdk.operator.api.reconciler.dependent.DependentResource;
+
+public @interface Dependent {
+
+  Class<?> resourceType();
+
+  Class<? extends DependentResource> type();
+}
