+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/config/dependent/DependentResourceConfiguration.java
@@ -1,6 +1,7 @@
-package io.javaoperatorsdk.operator.api.config;
+package io.javaoperatorsdk.operator.api.config.dependent;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.javaoperatorsdk.operator.api.reconciler.dependent.DependentResource;
 
 public interface DependentResourceConfiguration<R, P extends HasMetadata> {
