+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/dependent/DependentResource.java
@@ -1,6 +1,7 @@
-package io.javaoperatorsdk.operator.api.config;
+package io.javaoperatorsdk.operator.api.reconciler.dependent;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.javaoperatorsdk.operator.api.config.Utils;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.api.reconciler.EventSourceContext;
 import io.javaoperatorsdk.operator.processing.event.source.EventSource;
@@ -8,7 +9,7 @@ import io.javaoperatorsdk.operator.processing.event.source.EventSource;
 public interface DependentResource<R, P extends HasMetadata> {
   default EventSource initEventSource(EventSourceContext<P> context) {
     throw new IllegalStateException("Must be implemented if not automatically provided by the SDK");
-  };
+  }
 
   default Class<R> resourceType() {
     return (Class<R>) Utils.getFirstTypeArgumentFromInterface(getClass());
